task2: add youtube video url

+++ Task2/arkpz-pzpi-22-6-Aranzhyi-Ruslan-task2/idk/schemas/forecast.py
@@ -3,5 +3,5 @@ from pydantic import BaseModel
 
 class ForecastResponse(BaseModel):
     info_text: str
-    details: dict[str, int | str]
+    details: dict[str, float | str]
     # TODO: add other fields, like temperature array for next couple of days, etc.
